+++ example/src/stories/gestures/api/rotate/GestureRotateBasic.stories.ts
@@ -1,0 +1,13 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import Example from './GestureRotateBasic';
+
+const meta = {
+  title: 'Gestures/Rotate/Basic',
+  component: Example,
+} satisfies Meta<typeof Example>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {};